casa_test_support

Module gridder_interop

C++ casacore ConvolveGridder interop helpers.

Structs

GridderImage2d

Result of making a corrected dirty image with C++ casacore gridding helpers.

GridderPredictedVisibility

Result of degridding one visibility from a model image with C++ casacore.

GridderSampleCell

One nonzero cell written by C++ ConvolveGridder::grid() for a unit sample.

GridderSamplePatch

Result of gridding a single unit sample with C++ ConvolveGridder.

Functions

cpp_convolve_gridder_correction_row_2d

Return a 1D correction row from C++ casacore ConvolveGridder::correctX1D().

cpp_convolve_gridder_grid_unit_sample_2d

Grid a single unit sample with C++ casacore ConvolveGridder.

cpp_convolve_gridder_make_dirty_image_2d

Make a corrected dirty image with C++ casacore ConvolveGridder + LatticeFFT.

cpp_convolve_gridder_make_model_residual_image_2d

Make a corrected residual image from visibilities and a model image with C++ casacore ConvolveGridder + LatticeFFT.

cpp_convolve_gridder_predict_visibility_2d

Predict one visibility from a model image with C++ casacore ConvolveGridder.