casa_test_support::taql_interop

Function rust_taql_exec_ok

pub fn rust_taql_exec_ok(table: &mut Table, query: &str) -> Result<(), String>

Execute a TaQL statement via execute_taql(), returning Ok(()) on success.

Use this for queries that produce computed columns (GROUP BY, expressions in SELECT, etc.) where table.query() would fail because the result is not a simple RefTable.